dwave/requirements.txt
annotated-types==0.7.0
Authlib==1.3.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dimod==0.12.17
diskcache==5.6.3
dwave-cloud-client==0.13.1
dwave-gate==0.3.2
dwave-greedy==0.3.0
dwave-hybrid==0.6.12
dwave-inspector==0.5.1
dwave-neal==0.6.0
dwave-ocean-sdk==8.0.1
dwave-optimization==0.3.0
dwave-preprocessing==0.6.6
dwave-samplers==1.3.0
dwave-system==1.26.0
dwave-tabu==0.5.0
dwave_networkx==0.8.15
dwavebinarycsp==0.3.0
fasteners==0.19
Flask==3.1.0
fonttools==4.55.0
homebase==1.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
minorminer==0.2.15
networkx==3.4.2
numpy==2.1.3
orjson==3.10.12
packaging==24.2
penaltymodel==1.1.0
pillow==11.0.0
plucky==0.4.3
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
